feat: add pydantic_settings for environment variables
This commit adds the pydantic_settings package to manage environment
variables. The package is used to create a Settings class that inherits
from BaseSettings. The Settings class contains the environment variables
used in the application. The environment variables are loaded from a
.env file using the env_file attribute in the Config class. The Settings
class is then used to access the environment variables in the
application.